[KHJ] Link 엔티티 내 자료형 변경 (int → long)

+++ ptis-core/src/main/java/com/thewayhome/ptis/core/entity/Link.java
@@ -41,6 +41,6 @@ public class Link extends BaseEntity {
     private Node edNode;
 
     @Column(name="cost", nullable = false)
-    private int cost;
+    private long cost;
 
 }
